refactor: rename example for sharing policies across threads and remove obsolete file
