+++ Examples.md
@@ -205,7 +205,7 @@ assert!(!html.contains(r#"p { border-bottom: 2px solid black; }"#));
 
 
 <details>
-<summary><b>Sharing A `Policy` Between Threads</b></summary>
+<summary><b>Sharing A `Policy` Across Threads</b></summary>
 
 ```rust
 use std::sync::Arc;
@@ -213,29 +213,26 @@ use std::sync::Arc;
 use dom_sanitizer::preset::table_policy;
 use dom_sanitizer::DenyAllPolicy;
 
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let policy = DenyAllPolicy::builder()
-        // Allow table elements
-        .merge(table_policy())
-        .remove_elements(&["style"])
-        // `html`, `head`, and `body` are always kept
-        .build();
-        
-    let shared_policy = Arc::new(policy);
+let policy = DenyAllPolicy::builder()
+    // Allow table elements
+    .merge(table_policy())
+    .remove_elements(&["style"])
+    // `html`, `head`, and `body` are always kept
+    .build();
+    
+let shared_policy = Arc::new(policy);
 
-    for _ in 0..4 {
-        let policy = shared_policy.clone();
-        std::thread::spawn(move || {
-            let contents: &str = include_str!("../test-pages/table.html");
-            let doc = dom_query::Document::from(contents);
-            policy.sanitize_document(&doc);
-            assert!(doc.select("table > tr > td").exists());
-            assert!(!doc.select("style").exists());
-        });
-    }
-
-    Ok(())
+for _ in 0..4 {
+    let policy = shared_policy.clone();
+    std::thread::spawn(move || {
+        let contents: &str = include_str!("../test-pages/table.html");
+        let doc = dom_query::Document::from(contents);
+        policy.sanitize_document(&doc);
+        assert!(doc.select("table > tr > td").exists());
+        assert!(!doc.select("style").exists());
+    });
 }
+
 ```
 </details>
 
@@ -486,5 +483,65 @@ policy.sanitize_document(&doc);
 assert!(!doc.select("div.ad-block").exists());
 assert_eq!(doc.select("div").length(), 2);
 assert_eq!(doc.select("p").length(), 2);
+```
+</details>
+
+
+<details>
+<summary><b>Sharing A `PluginPolicy` across Threads (atomic)</b></summary>
+*This example requires `atomic` feature.*
+
+This example demonstrates how to safely share and use a `PluginPolicy` across multiple threads. 
+It utilizes the `atomic` feature, which is required to share `dom_query::Document`.
+
+```rust
+#[cfg(feature = "atomic")]
+{
+    use std::sync::Arc;
+    use std::sync::mpsc::channel;
+
+    use html5ever::local_name;
+
+    use dom_sanitizer::plugin_policy::preset;
+    use dom_sanitizer::plugin_policy::PluginPolicy;
+    use dom_sanitizer::Restrictive;
+
+
+    let policy: PluginPolicy<Restrictive> = PluginPolicy::builder()
+        // Allow table elements
+        .exclude(preset::MatchLocalNames(vec![
+            local_name!("table"),
+            local_name!("tr"),
+            local_name!("th"),
+            local_name!("td"),
+        ]))
+        .remove(preset::MatchLocalName(local_name!("style")))
+        // `html`, `head`, and `body` are always kept
+        .build();
+        
+    dbg!(&policy);
+    let shared_policy = Arc::new(policy);
+
+    let (tx, rx) = channel();
+
+    for _ in 0..4 {
+        let policy = shared_policy.clone();
+        let thread_tx = tx.clone();
+        std::thread::spawn(move || {
+            let contents: &str = include_str!("../test-pages/table.html");
+            let doc = dom_query::Document::from(contents);
+            policy.sanitize_document(&doc);
+            thread_tx.send(doc).unwrap();
+            
+        });
+        
+    }
+    drop(tx);
+
+    for doc in rx {
+        assert!(!doc.select("style").exists());
+        assert!(doc.select("table > tr > td").exists());
+    }
+}
 ```
 </details>
