+++ examples/shared_plugin_policy.rs
@@ -1,10 +1,12 @@
 use std::sync::Arc;
+use std::sync::mpsc::channel;
+
+use html5ever::local_name;
 
 use dom_sanitizer::plugin_policy::preset;
 use dom_sanitizer::plugin_policy::PluginPolicy;
 use dom_sanitizer::Restrictive;
 
-use html5ever::local_name;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let policy: PluginPolicy<Restrictive> = PluginPolicy::builder()
@@ -19,17 +21,28 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         // `html`, `head`, and `body` are always kept
         .build();
         
+    dbg!(&policy);
     let shared_policy = Arc::new(policy);
+
+    let (tx, rx) = channel();
 
     for _ in 0..4 {
         let policy = shared_policy.clone();
+        let thread_tx = tx.clone();
         std::thread::spawn(move || {
             let contents: &str = include_str!("../test-pages/table.html");
             let doc = dom_query::Document::from(contents);
             policy.sanitize_document(&doc);
-            assert!(doc.select("table > tr > td").exists());
-            assert!(!doc.select("style").exists());
+            thread_tx.send(doc).unwrap();
+            
         });
+        
+    }
+    drop(tx);
+
+    for doc in rx {
+        assert!(!doc.select("style").exists());
+        assert!(doc.select("table > tr > td").exists());
     }
 
     Ok(())
